(improvement)(Headless) Refactor ChatLayerService and SemanticLayerService
