+++ headless/chat/src/main/java/com/tencent/supersonic/headless/chat/query/llm/s2sql/LLMSqlQuery.java
@@ -1,12 +1,12 @@
 package com.tencent.supersonic.headless.chat.query.llm.s2sql;
 
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
-import com.tencent.supersonic.headless.api.pojo.SemanticSchema;
+import com.tencent.supersonic.headless.api.pojo.DataSetSchema;
 import com.tencent.supersonic.headless.api.pojo.SqlInfo;
 import com.tencent.supersonic.headless.api.pojo.request.SemanticQueryReq;
-import com.tencent.supersonic.headless.chat.utils.QueryReqBuilder;
 import com.tencent.supersonic.headless.chat.query.QueryManager;
 import com.tencent.supersonic.headless.chat.query.llm.LLMSemanticQuery;
+import com.tencent.supersonic.headless.chat.utils.QueryReqBuilder;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 
@@ -31,7 +31,7 @@ public class LLMSqlQuery extends LLMSemanticQuery {
     }
 
     @Override
-    public void initS2Sql(SemanticSchema semanticSchema, User user) {
+    public void initS2Sql(DataSetSchema dataSetSchema, User user) {
         SqlInfo sqlInfo = parseInfo.getSqlInfo();
         sqlInfo.setCorrectedS2SQL(sqlInfo.getParsedS2SQL());
     }
