+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/service/ChatLayerService.java
@@ -2,15 +2,12 @@ package com.tencent.supersonic.headless.server.facade.service;
 
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
 import com.tencent.supersonic.headless.api.pojo.SqlEvaluation;
-import com.tencent.supersonic.headless.api.pojo.request.DimensionValueReq;
-import com.tencent.supersonic.headless.api.pojo.request.QueryDataReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryMapReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryNLReq;
 import com.tencent.supersonic.headless.api.pojo.request.QuerySqlReq;
 import com.tencent.supersonic.headless.api.pojo.response.MapInfoResp;
 import com.tencent.supersonic.headless.api.pojo.response.MapResp;
 import com.tencent.supersonic.headless.api.pojo.response.ParseResp;
-import com.tencent.supersonic.headless.api.pojo.response.QueryResult;
 
 /***dd
  * SemanticLayerService for query and search
@@ -20,10 +17,6 @@ public interface ChatLayerService {
     MapResp performMapping(QueryNLReq queryNLReq);
 
     ParseResp performParsing(QueryNLReq queryNLReq);
-
-    QueryResult executeDirectQuery(QueryDataReq queryData, User user) throws Exception;
-
-    Object queryDimensionValue(DimensionValueReq dimensionValueReq, User user) throws Exception;
 
     MapInfoResp map(QueryMapReq queryMapReq);
