+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/service/SemanticLayerService.java
@@ -4,11 +4,11 @@ import com.tencent.supersonic.auth.api.authentication.pojo.User;
 import com.tencent.supersonic.headless.api.pojo.DataSetSchema;
 import com.tencent.supersonic.headless.api.pojo.EntityInfo;
 import com.tencent.supersonic.headless.api.pojo.SemanticParseInfo;
-import com.tencent.supersonic.headless.api.pojo.request.QueryDimValueReq;
+import com.tencent.supersonic.headless.api.pojo.request.DimensionValueReq;
 import com.tencent.supersonic.headless.api.pojo.request.SemanticQueryReq;
-import com.tencent.supersonic.headless.api.pojo.response.SemanticTranslateResp;
 import com.tencent.supersonic.headless.api.pojo.response.ItemResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticQueryResp;
+import com.tencent.supersonic.headless.api.pojo.response.SemanticTranslateResp;
 
 import java.util.List;
 
@@ -23,7 +23,7 @@ public interface SemanticLayerService {
 
     SemanticQueryResp queryByReq(SemanticQueryReq queryReq, User user) throws Exception;
 
-    SemanticQueryResp queryDimValue(QueryDimValueReq queryDimValueReq, User user);
+    SemanticQueryResp queryDimensionValue(DimensionValueReq dimensionValueReq, User user);
 
     EntityInfo getEntityInfo(SemanticParseInfo parseInfo, DataSetSchema dataSetSchema, User user);
