+++ headless/server/src/main/java/com/tencent/supersonic/headless/server/facade/service/impl/S2SemanticLayerService.java
@@ -4,6 +4,7 @@ import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
 import com.tencent.supersonic.common.pojo.DateConf;
+import com.tencent.supersonic.common.pojo.QueryColumn;
 import com.tencent.supersonic.common.pojo.enums.FilterOperatorEnum;
 import com.tencent.supersonic.common.pojo.enums.QueryType;
 import com.tencent.supersonic.common.pojo.enums.TaskStatusEnum;
@@ -19,7 +20,8 @@ import com.tencent.supersonic.headless.api.pojo.SchemaElementType;
 import com.tencent.supersonic.headless.api.pojo.SemanticParseInfo;
 import com.tencent.supersonic.headless.api.pojo.TagTypeDefaultConfig;
 import com.tencent.supersonic.headless.api.pojo.TimeDefaultConfig;
-import com.tencent.supersonic.headless.api.pojo.request.QueryDimValueReq;
+import com.tencent.supersonic.headless.api.pojo.enums.SemanticType;
+import com.tencent.supersonic.headless.api.pojo.request.DimensionValueReq;
 import com.tencent.supersonic.headless.api.pojo.request.QueryFilter;
 import com.tencent.supersonic.headless.api.pojo.request.QueryMultiStructReq;
 import com.tencent.supersonic.headless.api.pojo.request.QuerySqlReq;
@@ -27,11 +29,17 @@ import com.tencent.supersonic.headless.api.pojo.request.QueryStructReq;
 import com.tencent.supersonic.headless.api.pojo.request.SchemaFilterReq;
 import com.tencent.supersonic.headless.api.pojo.request.SemanticQueryReq;
 import com.tencent.supersonic.headless.api.pojo.response.DimensionResp;
-import com.tencent.supersonic.headless.api.pojo.response.SemanticTranslateResp;
 import com.tencent.supersonic.headless.api.pojo.response.ItemResp;
 import com.tencent.supersonic.headless.api.pojo.response.ModelResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticQueryResp;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticSchemaResp;
+import com.tencent.supersonic.headless.api.pojo.response.SemanticTranslateResp;
+import com.tencent.supersonic.headless.chat.knowledge.HanlpMapResult;
+import com.tencent.supersonic.headless.chat.knowledge.KnowledgeBaseService;
+import com.tencent.supersonic.headless.chat.knowledge.MapResult;
+import com.tencent.supersonic.headless.chat.knowledge.SearchService;
+import com.tencent.supersonic.headless.chat.knowledge.helper.HanlpHelper;
+import com.tencent.supersonic.headless.chat.knowledge.helper.NatureHelper;
 import com.tencent.supersonic.headless.chat.utils.QueryReqBuilder;
 import com.tencent.supersonic.headless.core.cache.QueryCache;
 import com.tencent.supersonic.headless.core.executor.QueryExecutor;
@@ -57,6 +65,7 @@ import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.Map;
@@ -77,6 +86,7 @@ public class S2SemanticLayerService implements SemanticLayerService {
     private final SchemaService schemaService;
     private final SemanticTranslator semanticTranslator;
     private final MetricDrillDownChecker metricDrillDownChecker;
+    private final KnowledgeBaseService knowledgeBaseService;
     private QueryCache queryCache = ComponentFactory.getQueryCache();
     private List<QueryExecutor> queryExecutors = ComponentFactory.getQueryExecutors();
 
@@ -88,7 +98,8 @@ public class S2SemanticLayerService implements SemanticLayerService {
             DataSetService dataSetService,
             SchemaService schemaService,
             SemanticTranslator semanticTranslator,
-            MetricDrillDownChecker metricDrillDownChecker) {
+            MetricDrillDownChecker metricDrillDownChecker,
+            KnowledgeBaseService knowledgeBaseService) {
         this.statUtils = statUtils;
         this.queryUtils = queryUtils;
         this.queryReqConverter = queryReqConverter;
@@ -97,6 +108,7 @@ public class S2SemanticLayerService implements SemanticLayerService {
         this.schemaService = schemaService;
         this.semanticTranslator = semanticTranslator;
         this.metricDrillDownChecker = metricDrillDownChecker;
+        this.knowledgeBaseService = knowledgeBaseService;
     }
 
     public DataSetSchema getDataSetSchema(Long id) {
@@ -175,10 +187,72 @@ public class S2SemanticLayerService implements SemanticLayerService {
     }
 
     @Override
-    @SneakyThrows
-    public SemanticQueryResp queryDimValue(QueryDimValueReq queryDimValueReq, User user) {
-        QuerySqlReq querySqlReq = buildQuerySqlReq(queryDimValueReq);
+    public SemanticQueryResp queryDimensionValue(DimensionValueReq dimensionValueReq, User user) {
+        SemanticQueryResp semanticQueryResp = new SemanticQueryResp();
+        DimensionResp dimensionResp = getDimension(dimensionValueReq);
+        Set<Long> dataSetIds = dimensionValueReq.getDataSetIds();
+        dimensionValueReq.setModelId(dimensionResp.getModelId());
+        List<String> dimensionValues = getDimensionValuesFromDict(dimensionValueReq, dataSetIds);
+        // if the search results is null,search dimensionValue from database
+        if (CollectionUtils.isEmpty(dimensionValues)) {
+            semanticQueryResp = getDimensionValuesFromDb(dimensionValueReq, user);
+            return semanticQueryResp;
+        }
+        List<QueryColumn> columns = new ArrayList<>();
+        QueryColumn queryColumn = new QueryColumn();
+        queryColumn.setNameEn(dimensionValueReq.getBizName());
+        queryColumn.setShowType(SemanticType.CATEGORY.name());
+        queryColumn.setAuthorized(true);
+        queryColumn.setType("CHAR");
+        columns.add(queryColumn);
+        List<Map<String, Object>> resultList = new ArrayList<>();
+        dimensionValues.stream().forEach(o -> {
+            Map<String, Object> map = new HashMap<>();
+            map.put(dimensionValueReq.getBizName(), o);
+            resultList.add(map);
+        });
+        semanticQueryResp.setColumns(columns);
+        semanticQueryResp.setResultList(resultList);
+        return semanticQueryResp;
+    }
+
+    private List<String> getDimensionValuesFromDict(DimensionValueReq dimensionValueReq, Set<Long> dataSetIds) {
+        //if value is null ,then search from NATURE_TO_VALUES
+        if (StringUtils.isBlank(dimensionValueReq.getValue())) {
+            return SearchService.getDimensionValue(dimensionValueReq);
+        }
+        Map<Long, List<Long>> modelIdToDataSetIds = new HashMap<>();
+        modelIdToDataSetIds.put(dimensionValueReq.getModelId(), new ArrayList<>(dataSetIds));
+        //search from prefixSearch
+        List<HanlpMapResult> hanlpMapResultList = knowledgeBaseService.prefixSearch(dimensionValueReq.getValue(),
+                2000, modelIdToDataSetIds, dataSetIds);
+        HanlpHelper.transLetterOriginal(hanlpMapResultList);
+        return hanlpMapResultList.stream()
+                .filter(o -> {
+                    for (String nature : o.getNatures()) {
+                        Long elementID = NatureHelper.getElementID(nature);
+                        if (dimensionValueReq.getElementID().equals(elementID)) {
+                            return true;
+                        }
+                    }
+                    return false;
+                })
+                .map(MapResult::getName)
+                .collect(Collectors.toList());
+    }
+
+    private SemanticQueryResp getDimensionValuesFromDb(DimensionValueReq dimensionValueReq, User user) {
+        QuerySqlReq querySqlReq = buildQuerySqlReq(dimensionValueReq);
         return queryByReq(querySqlReq, user);
+    }
+
+    private DimensionResp getDimension(DimensionValueReq dimensionValueReq) {
+        DimensionResp dimensionResp = schemaService.getDimension(dimensionValueReq.getElementID());
+        if (dimensionResp == null) {
+            return schemaService.getDimension(dimensionValueReq.getBizName(),
+                    dimensionValueReq.getModelId());
+        }
+        return dimensionResp;
     }
 
     public EntityInfo getEntityInfo(SemanticParseInfo parseInfo, DataSetSchema dataSetSchema, User user) {
@@ -291,10 +365,10 @@ public class S2SemanticLayerService implements SemanticLayerService {
         return schemaFilterReq;
     }
 
-    private QuerySqlReq buildQuerySqlReq(QueryDimValueReq queryDimValueReq) {
+    private QuerySqlReq buildQuerySqlReq(DimensionValueReq queryDimValueReq) {
         QuerySqlReq querySqlReq = new QuerySqlReq();
         List<ModelResp> modelResps = schemaService.getModelList(Lists.newArrayList(queryDimValueReq.getModelId()));
-        DimensionResp dimensionResp = schemaService.getDimension(queryDimValueReq.getDimensionBizName(),
+        DimensionResp dimensionResp = schemaService.getDimension(queryDimValueReq.getBizName(),
                 queryDimValueReq.getModelId());
         ModelResp modelResp = modelResps.get(0);
         String sql = String.format("select distinct %s from %s where 1=1",
@@ -306,7 +380,7 @@ public class S2SemanticLayerService implements SemanticLayerService {
                     queryDimValueReq.getDateInfo().getEndDate());
         }
         if (StringUtils.isNotBlank(queryDimValueReq.getValue())) {
-            sql += " AND " + queryDimValueReq.getDimensionBizName() + " LIKE '%" + queryDimValueReq.getValue() + "%'";
+            sql += " AND " + queryDimValueReq.getBizName() + " LIKE '%" + queryDimValueReq.getValue() + "%'";
         }
         querySqlReq.setModelIds(Sets.newHashSet(queryDimValueReq.getModelId()));
         querySqlReq.setSql(sql);
