+++ launchers/standalone/src/test/java/com/tencent/supersonic/headless/QueryDimensionTest.java
@@ -1,7 +1,7 @@
 package com.tencent.supersonic.headless;
 
 import com.tencent.supersonic.auth.api.authentication.pojo.User;
-import com.tencent.supersonic.headless.api.pojo.request.QueryDimValueReq;
+import com.tencent.supersonic.headless.api.pojo.request.DimensionValueReq;
 import com.tencent.supersonic.headless.api.pojo.response.SemanticQueryResp;
 import org.junit.Assert;
 import org.junit.jupiter.api.Test;
@@ -10,11 +10,11 @@ public class QueryDimensionTest extends BaseTest {
 
     @Test
     public void testQueryDimValue() {
-        QueryDimValueReq queryDimValueReq = new QueryDimValueReq();
+        DimensionValueReq queryDimValueReq = new DimensionValueReq();
         queryDimValueReq.setModelId(1L);
-        queryDimValueReq.setDimensionBizName("department");
+        queryDimValueReq.setBizName("department");
 
-        SemanticQueryResp queryResp = semanticLayerService.queryDimValue(queryDimValueReq, User.getFakeUser());
+        SemanticQueryResp queryResp = semanticLayerService.queryDimensionValue(queryDimValueReq, User.getFakeUser());
         Assert.assertNotNull(queryResp.getResultList());
         Assert.assertEquals(4, queryResp.getResultList().size());
     }
